refactor(frontend): standardize error boundary usage across route components (#776)
Extract the duplicated Sentry-reporting/retry/fallback logic from the
three copy-pasted route error.tsx files into a shared RouteError
component, and add previously missing error boundaries for the
[locale] and admin route segments.
